# Optimization for geom_local_to_global