Currently, the function geom_local_to_global is slowing down the apptronik apollo benchmark because we have a lot of geom composing the ground.
We recalculate at each timestep their position while it is unnecessary.
This change aim to skip unnecessary calculation.

Benchmark on a L40:
`python mujoco_warp/testspeed.py --function=step --mjcf=/home/horde/Documents/optimization/mujoco_warp/benchmark/apptronik_apollo/scene_terrain.xml --batch_size=8192 --nstep=1000`

Trunk:
```
 Total JIT time: 1.40 s
 Total simulation time: 2.39 s
 Total steps per second: 342,123
 Total realtime factor: 1,710.61 x
 Total time per step: 2922.93 ns
```

This branch:
```
 Total JIT time: 1.56 s
 Total simulation time: 20.52 s
 Total steps per second: 399,152
 Total realtime factor: 1,995.76 x
 Total time per step: 2505.31 ns
```